refactor(preview): extract pan logic into useKeyboardPan hook
preview.tsx was getting long for a single component. Move the
keyboard handler (and its isEditableTarget helper) into a dedicated
hook with its own test file. The KeyboardPan component shrinks to a
one-liner that just hosts the hook inside TransformWrapper.
