+++ frontend/src/components/preview.test.tsx
@@ -6,10 +6,12 @@ import { setupRender } from "../test/render";
 const mockZoomIn = vi.fn();
 const mockZoomOut = vi.fn();
 const mockResetTransform = vi.fn();
-const mockSetTransformState = vi.fn();
 
 let mockScale = 1;
-const mockTransformState = { scale: 1, positionX: 0, positionY: 0 };
+
+vi.mock("./use-keyboard-pan", () => ({
+  useKeyboardPan: () => {},
+}));
 
 vi.mock("react-zoom-pan-pinch", () => ({
   TransformWrapper: ({ children }: { children: ReactNode }) => children,
@@ -21,10 +23,6 @@ vi.mock("react-zoom-pan-pinch", () => ({
   }),
   useTransformComponent: (cb: (s: { state: { scale: number } }) => unknown) =>
     cb({ state: { scale: mockScale } }),
-  useTransformContext: () => ({
-    transformState: mockTransformState,
-    setTransformState: mockSetTransformState,
-  }),
 }));
 
 const SAMPLE_SVG = "data:image/png;base64,AAAA";
@@ -35,9 +33,6 @@ const render = setupRender();
 
 beforeEach(() => {
   mockScale = 1;
-  mockTransformState.scale = 1;
-  mockTransformState.positionX = 100;
-  mockTransformState.positionY = 200;
   vi.clearAllMocks();
 });
 
@@ -102,58 +97,6 @@ describe("Preview", () => {
       mockScale = 1;
       render(<Preview svg="data:image/png;base64,BBBB" />);
       expect(zoomLevel()).toBe("100%");
-    });
-  });
-
-  describe("arrow key panning", () => {
-    const pressKey = (key: string, init: KeyboardEventInit = {}) => {
-      act(() => {
-        window.dispatchEvent(new KeyboardEvent("keydown", { key, bubbles: true, ...init }));
-      });
-    };
-
-    it.each([
-      { key: "ArrowLeft", dx: 50, dy: 0 },
-      { key: "ArrowRight", dx: -50, dy: 0 },
-      { key: "ArrowUp", dx: 0, dy: 50 },
-      { key: "ArrowDown", dx: 0, dy: -50 },
-    ])("$key shifts the transform by ($dx, $dy)", ({ key, dx, dy }) => {
-      render(<Preview svg={SAMPLE_SVG} />);
-      pressKey(key);
-      expect(mockSetTransformState).toHaveBeenCalledWith(1, 100 + dx, 200 + dy);
-    });
-
-    it("uses a larger step when Shift is held", () => {
-      render(<Preview svg={SAMPLE_SVG} />);
-      pressKey("ArrowRight", { shiftKey: true });
-      expect(mockSetTransformState).toHaveBeenCalledWith(1, 100 - 200, 200);
-    });
-
-    it("ignores non-arrow keys", () => {
-      render(<Preview svg={SAMPLE_SVG} />);
-      pressKey("a");
-      pressKey("Enter");
-      expect(mockSetTransformState).not.toHaveBeenCalled();
-    });
-
-    it.each(["INPUT", "TEXTAREA"])("ignores arrows while typing in a %s", (tagName) => {
-      render(<Preview svg={SAMPLE_SVG} />);
-      const editable = document.createElement(tagName.toLowerCase()) as HTMLElement;
-      document.body.appendChild(editable);
-      editable.focus();
-      act(() => {
-        editable.dispatchEvent(new KeyboardEvent("keydown", { key: "ArrowLeft", bubbles: true }));
-      });
-      expect(mockSetTransformState).not.toHaveBeenCalled();
-      document.body.removeChild(editable);
-    });
-
-    it("reads the latest position on each keypress", () => {
-      render(<Preview svg={SAMPLE_SVG} />);
-      pressKey("ArrowRight");
-      mockTransformState.positionX = 999;
-      pressKey("ArrowRight");
-      expect(mockSetTransformState).toHaveBeenNthCalledWith(2, 1, 999 - 50, 200);
     });
   });
 });
